feat: Add DisplayUtils Class for displaying cars status
